+++ src/components/Dropdown.tsx
@@ -1,4 +1,4 @@
-import { ChevronDownIcon } from './icons';
+import { ChevronDown } from "lucide-react";
 
 type DropdownProps = {
   options: { label: string; value: string }[];
@@ -32,7 +32,7 @@ const Dropdown = ({
         className="pointer-events-none absolute right-2 top-1/2 -translate-y-1/2 rounded-lg border border-grey-1100 bg-containerL3 bg-gradient-to-b 
      from-[rgba(102,102,102,0.06)] to-[rgba(0,0,0,0.06)] p-1 shadow-inner  shadow-grey-600"
       >
-        <ChevronDownIcon className="pointer-events-none size-4  text-mid" />
+        <ChevronDown className="pointer-events-none size-4  text-mid" />
       </div>
     </div>
   );