+++ src/components/GlobalDataProvider.tsx
@@ -1,8 +1,8 @@
-import { AO_CU_URL, log } from '@src/constants';
-import { useEffectOnce } from '@src/hooks/useEffectOnce';
+import { log } from '@src/constants';
 import { useGlobalState } from '@src/store';
 import { cleanupDbCache } from '@src/store/db';
 import { showErrorToast } from '@src/utils/toast';
+import { useQueryClient } from '@tanstack/react-query';
 import { ReactElement, useEffect } from 'react';
 
 // Time to wait in ms to check if the AO CU URL is congested
@@ -12,13 +12,20 @@ const TWO_MINUTES = 120000;
 const GlobalDataProvider = ({ children }: { children: ReactElement }) => {
   const setBlockHeight = useGlobalState((state) => state.setBlockHeight);
   const setCurrentEpoch = useGlobalState((state) => state.setCurrentEpoch);
+  const currentEpoch = useGlobalState((state) => state.currentEpoch);
   const setTicker = useGlobalState((state) => state.setTicker);
   const arweave = useGlobalState((state) => state.arweave);
   const arioReadSDK = useGlobalState((state) => state.arIOReadSDK);
   const setAoCongested = useGlobalState((state) => state.setAoCongested);
+  const arioProcessId = useGlobalState((state) => state.arioProcessId);
+  const aoCuUrl = useGlobalState((state) => state.aoCuUrl);
+  const networkPortalDB = useGlobalState((state) => state.networkPortalDB);
+  const queryClient = useQueryClient();
 
-  useEffectOnce(() => {
+  useEffect(() => {
     const update = async () => {
+      queryClient.resetQueries();
+
       // perform this first as retrieving the current epic takes some time
       const { Ticker } = await arioReadSDK.getInfo();
       setTicker(Ticker);
@@ -37,20 +44,22 @@ const GlobalDataProvider = ({ children }: { children: ReactElement }) => {
           return;
         }
         setCurrentEpoch(currentEpoch);
-
-        if (currentEpoch?.epochIndex) {
-          cleanupDbCache(currentEpoch.epochIndex);
-        }
       } catch (error) {
         log.error('Error fetching current epoch');
-          showErrorToast(
-            'Error fetching current epoch. Application may not function as expected.',
-          );
+        showErrorToast(
+          'Error fetching current epoch. Application may not function as expected.',
+        );
       }
     };
 
     update();
-  });
+  }, [arioReadSDK, queryClient, setCurrentEpoch, setTicker]);
+
+  useEffect(() => {
+    if (currentEpoch?.epochIndex && networkPortalDB) {
+      cleanupDbCache(networkPortalDB, currentEpoch.epochIndex);
+    }
+  }, [currentEpoch, networkPortalDB]);
 
   useEffect(() => {
     // Block Height Updater
@@ -66,7 +75,7 @@ const GlobalDataProvider = ({ children }: { children: ReactElement }) => {
 
     const checkAoCongestion = () => {
       const startTime = Date.now();
-      fetch(AO_CU_URL, { method: 'HEAD' })
+      fetch(`${aoCuUrl}/state/${arioProcessId}`, { method: 'HEAD' })
         .then((res) => {
           const endTime = Date.now();
           if (!res.ok) {
@@ -90,7 +99,7 @@ const GlobalDataProvider = ({ children }: { children: ReactElement }) => {
       clearInterval(congestionInterval);
       setAoCongested(false);
     };
-  });
+  }, [aoCuUrl, arioProcessId, arweave.blocks, setAoCongested, setBlockHeight]);
 
   return <>{children}</>;
 };