+++ src/components/icons/index.ts
@@ -7,9 +7,6 @@ import BinocularsGradientIcon from './binoculars_gradient.svg?react';
 import CancelButtonXIcon from './cancel_button_x.svg?react';
 import CaretDoubleRightIcon from './caret_double_right.svg?react';
 import CheckSquareIcon from './check_square.svg?react';
-import ChevronDownIcon from './chevron_down.svg?react';
-import CircleCheckIcon from './circle_check.svg?react';
-import CircleIcon from './circle.svg?react';
 import ClockRewindIcon from './clock_rewind.svg?react';
 import CloseIcon from './close.svg?react';
 import CloseDrawerIcon from './close_drawer.svg?react';
@@ -19,15 +16,12 @@ import CopyIcon from './copy.svg?react';
 import CopyCheckedIcon from './copy_checked.svg?react';
 import DashboardIcon from './dashboard.svg?react';
 import DocsIcon from './docs.svg?react';
-import DownloadIcon from './download.svg?react';
 import EditIcon from './edit_icon.svg?react';
 import FailSquareIcon from './fail_square.svg?react';
 import FormErrorIcon from './form_error.svg?react';
 import GatewayIcon from './gateway.svg?react';
 import GatewayHoverIcon from './gateway_hover.svg?react';
 import GatewaysIcon from './gateways.svg?react';
-import GearIcon from './gear.svg?react';
-import HeaderSeparatorIcon from './header_separator.svg?react';
 import InstantWithdrawalIcon from './instant_withdrawal.svg?react';
 import InfoIcon from './info_icon.svg?react';
 import LinkArrowIcon from './link_arrow.svg?react';
@@ -37,7 +31,6 @@ import ObserversConnectIcon from './observers_connect_icon.svg?react';
 import OpenDrawerIcon from './open_drawer.svg?react';
 import PinkArrowIcon from './pink_arrow.svg?react';
 import ResetIcon from './reset.svg?react';
-import ReportsIcon from './reports.svg?react';
 import SortAsc from './sort_asc.svg?react';
 import SortDesc from './sort_desc.svg?react';
 import StakingIcon from './staking.svg?react';
@@ -48,9 +41,7 @@ import StreakDownArrowIcon from './streak_down_arrow.svg?react';
 import StreakUpArrowIcon from './streak_up_arrow.svg?react';
 import SuccessCheck from './success_check.svg?react';
 import ThreeDotsIcon from './three_dots_icon.svg?react';
-import TimerIcon from './timer.svg?react';
 import ToastCloseIcon from './toast_close.svg?react';
-import WalletIcon from './wallet.svg?react';
 import WarningIcon from './warning.svg?react';
 import WarningTriangleIcon from './warning_triangle.svg?react';
 
@@ -63,9 +54,6 @@ export {
   CaretDoubleRightIcon,
   CancelButtonXIcon,
   CheckSquareIcon,
-  ChevronDownIcon,
-  CircleCheckIcon,
-  CircleIcon,
   ClockRewindIcon,
   CloseDrawerIcon,
   CloseIcon,
@@ -75,15 +63,12 @@ export {
   CopyIcon,
   DashboardIcon,
   DocsIcon,
-  DownloadIcon,
   EditIcon,
   FailSquareIcon,
   FormErrorIcon,
   GatewayHoverIcon,
   GatewayIcon,
   GatewaysIcon,
-  GearIcon,
-  HeaderSeparatorIcon,
   InfoIcon,
   InstantWithdrawalIcon,
   LinkArrowIcon,
@@ -93,7 +78,6 @@ export {
   OpenDrawerIcon,
   PinkArrowIcon,
   ResetIcon,
-  ReportsIcon,
   SortAsc,
   SortDesc,
   StakingIcon,
@@ -104,9 +88,7 @@ export {
   StreakUpArrowIcon,
   SuccessCheck,
   ThreeDotsIcon,
-  TimerIcon,
   ToastCloseIcon,
-  WalletIcon,
   WarningIcon,
   WarningTriangleIcon,
 };