+++ src/hooks/useEpochs.ts
@@ -2,10 +2,14 @@ import { useGlobalState } from '@src/store';
 import { getEpoch } from '@src/store/db';
 import { useQuery } from '@tanstack/react-query';
 
-/** Returns last 14 epochs */
+// TODO: this could be a parameter provided by the user now
+const HISTORICAL_EPOCHS_TO_FETCH = 13;
+
+/** Returns last EPOCHS_TO_FETCH epochs */
 const useEpochs = () => {
   const arIOReadSDK = useGlobalState((state) => state.arIOReadSDK);
   const startEpoch = useGlobalState((state) => state.currentEpoch);
+  const networkPortalDB = useGlobalState((state) => state.networkPortalDB);
 
   const queryResults = useQuery({
     queryKey: ['epochs', arIOReadSDK, startEpoch],
@@ -15,11 +19,18 @@ const useEpochs = () => {
       }
 
       const additionalEpochs = await Promise.all(
-        Array.from({ length: 13 }, (_, index) => startEpoch.epochIndex - index - 1)
-          .map(epochIndex => getEpoch(arIOReadSDK, epochIndex))
-      );      
+        Array.from(
+          { length: HISTORICAL_EPOCHS_TO_FETCH },
+          (_, index) => startEpoch.epochIndex - index - 1,
+        ).map((epochIndex) =>
+          getEpoch(networkPortalDB, arIOReadSDK, epochIndex),
+        ),
+      );
 
-      return [startEpoch, ...additionalEpochs.filter(e => e !== undefined)];
+      return [
+        startEpoch,
+        ...additionalEpochs.filter((e: any) => e !== undefined),
+      ];
     },
   });