+++ src/hooks/usePrescribedNames.ts
@@ -9,7 +9,7 @@ const usePrescribedNames = () => {
   const currentEpoch = useGlobalState((state) => state.currentEpoch);
 
   const queryResults = useQuery({
-    queryKey: ['prescribedNames', currentEpoch?.epochIndex || -1],
+    queryKey: ['prescribedNames', arIOReadSDK, currentEpoch?.epochIndex || -1],
     queryFn: () => {
       if (arIOReadSDK && currentEpoch) {
         return arIOReadSDK.getPrescribedNames(currentEpoch).catch((e) => {